feat: use the workspace manifest root
When using `cargo locate-project`, specify the `--workspace` option.
This will provide a different result when within a workspace, providing
the path to the top-level `Cargo.toml` file instead of the one for the
workspace member. Patch changes to workspace member TOML files will not
alter builds and will display a warning instead.

For projects that do not use workspaces, the output of this command is
identical to the previous version.